fix: declare module "obsidian" 内类型未定义
