+++ src/main.ts
@@ -11,21 +11,6 @@ import '../style/suggest.css'
 import { AttachmentsModal } from "./ui/attachments/AttachmentsModal";
 import { getLocal } from "./i18/messages";
 
-declare module "obsidian" {
-	interface CanvasView extends TextFileView {
-		handlePaste: (e: ClipboardEvent) => Promise<void>;
-		/** Obsidian 未公开的 Canvas 内部对象，仅声明本插件用到的成员 */
-		canvas: {
-			posCenter(): { x: number; y: number };
-			createFileNode(options: {
-				pos: { x: number; y: number };
-				position: string;
-				file: TFile;
-			}): unknown;
-		};
-	}
-}
-
 export default class AttachmentProPlugin extends Plugin {
 	settings: AttachmentProConfig;
